Revert "Move server App to the client"
This reverts commit 1e18f8386c3426cd57c5b15995fca8c7e61adfa5.
