--- modules/ssr/client-react/index.tsx
@@ -1,4 +1,0 @@
-import ClientModule from '@gqlapp/module-client-react';
-import createServerApp from './createServerApp';
-
-export default new ClientModule({ createServerApp: [createServerApp] });
