--- modules/ssr/client-react/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@gqlapp/ssr-client-react",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@gqlapp/counter-common": "^0.1.0"
-  }
-}
