+++ modules/ssr/server-ts/index.ts
@@ -2,7 +2,7 @@ import path from 'path';
 import { Express, Request, Response } from 'express';
 import { SharedOptions } from '@gqlapp/module-server-ts';
 import SsrModule from './SsrModule';
-const reactRenderer = require('./react').default;
+import reactRenderer from './react';
 
 const renderServerSide = (sharedOptions: SharedOptions) => async (
   req: Request,
