+++ packages/client/src/modules.ts
@@ -8,11 +8,9 @@ import validation from '@gqlapp/validation-common-react';
 import ClientModule from '@gqlapp/module-client-react';
 import defaultRouter from '@gqlapp/router-client-react';
 import payments from '@gqlapp/payments-client-react';
-import ssr from '@gqlapp/ssr-client-react';
 import '@gqlapp/favicon-common';
 
 const post = require('@gqlapp/post-client-react').default;
-// const ssr = require('@gqlapp/ssr-client-react').default;
 const pageNotFound = require('@gqlapp/page-not-found-client-react').default;
 const reports = require('@gqlapp/reports-client-react').default;
 const upload = require('@gqlapp/upload-client-react').default;
@@ -34,7 +32,6 @@ const modules = new ClientModule(
   i18n,
   reports,
   pageNotFound,
-  ssr,
   core
 );
